remove user, view user-list, update user info. update user api

+++ banking/api/user.py
@@ -17,6 +17,8 @@ from django.http import JsonResponse, HttpResponse
 
 import urllib
 
+from django.db.models import Q
+
 class auth(APIView):
     def post(self, request, format=None):
         try:
@@ -82,18 +84,10 @@ class user(APIView):
             user = User.objects.get(pk=pk)
         elif pattern:
             pattern = urllib.unquote(pattern)
-            
-            '''
-                need to optimize to use one filter() by username, firstname and lastname
-            '''
-            users = User.objects.filter(username__startswith=pattern).distinct()
-            users = users | User.objects.filter(first_name__startswith=pattern).distinct()
-            users = users | User.objects.filter(last_name__startswith=pattern).distinct()
+            query = Q(username__startswith=pattern) | Q(first_name__startswith=pattern) | Q(last_name__startswith=pattern)
+            users = User.objects.filter(query).distinct()
             users = UserSerializer(users, many=True)
-            
-            return Response({
-                'users': users.data
-            })
+            return Response(users.data)
         else:    
             user = User.objects.get(auth_token=key)
         user = UserSerializer(user)
@@ -101,7 +95,7 @@ class user(APIView):
             'user': user.data
         })
         
-    def post(self, request, format=None):
+    def post(self, request, pk, pattern, format=None):
         try:
             data = request.data
         except ParseError as error:
@@ -132,16 +126,11 @@ class user(APIView):
         user.first_name = data['first_name']
         user.last_name = data['last_name']
         user.save()
-        users = User.objects.all()
+        users = User.objects.all().order_by('pk')
         users = UserSerializer(users, many=True)
-        return JsonResponse({
-            'users' : users.data
-        })
+        return Response(users.data)
         
-    def put(self, request, format=None):
-        pass
-    
-    def delete(self, request, format=None):
+    def put(self, request, pk, pattern, format=None):
         try:
             data = request.data
         except ParseError as error:
@@ -149,7 +138,7 @@ class user(APIView):
                 'Invalid JSON - {0}'.format(error.detail),
                 status=status.HTTP_400_BAD_REQUEST
             )
-        if 'username' not in data:
+        if 'username' not in data or 'password' not in data or 'first_name' not in data or 'last_name' not in data:
             return Response(
                 'Wrong credentials',
                 status=status.HTTP_401_UNAUTHORIZED
@@ -165,13 +154,44 @@ class user(APIView):
                 'Invalid HTTP request - {0}',
                 status=status.HTTP_400_BAD_REQUEST
             )
-        user = User.objects.get(username=data['username'])
-        user.delete();
-        users = User.objects.all()
+        user = User.objects.get(pk=pk)
+        if not user:
+            return Response(
+                'User does not exist',
+                status=status.HTTP_404_NOT_FOUND
+            )
+        user.username = data['username']
+        if data['password']:
+            user.password = data['password']
+        user.first_name = data['first_name']
+        user.last_name = data['last_name']
+        user.save()
+        users = User.objects.all().order_by('pk')
         users = UserSerializer(users, many=True)
-        return JsonResponse({
-            'users' : users.data
-        })
+        return Response(users.data)
+    
+    def delete(self, request, pk, pattern, format=None):
+        try:
+            if not has_permisions(request):
+                return HttpResponse(
+                    'You do not have permission',
+                    status=status.HTTP_403_FORBIDDEN
+                )
+        except ParseError:
+            return HttpResponse(
+                'Invalid HTTP request - {0}',
+                status=status.HTTP_400_BAD_REQUEST
+            )
+        user = User.objects.get(pk=pk)
+        if not user:
+            return Response(
+                'User does not exist',
+                status=status.HTTP_404_NOT_FOUND
+            )
+        user.delete();
+        users = User.objects.all().order_by('pk')
+        users = UserSerializer(users, many=True)
+        return Response(users.data)
         
 class user_list(APIView):
     authentication_classes = (
@@ -182,8 +202,6 @@ class user_list(APIView):
     )
     
     def get(self, request, format=None):
-        users = User.objects.all()
+        users = User.objects.all().order_by('pk')
         users = UserSerializer(users, many=True)
-        return JsonResponse({
-            'users' : users.data
-        })
+        return Response(users.data)
